Move SpecificEnthalpy out of EquationsOfState into Hydro
