+++ src/Evolution/Systems/GrMhd/ValenciaDivClean/PrimitiveFromConservative.cpp
@@ -14,7 +14,7 @@
 #include "Evolution/Systems/GrMhd/ValenciaDivClean/Tags.hpp"  // IWYU pragma: keep
 #include "PointwiseFunctions/GeneralRelativity/IndexManipulation.hpp"
 #include "PointwiseFunctions/GeneralRelativity/Tags.hpp"  // IWYU pragma: keep
-#include "PointwiseFunctions/Hydro/EquationsOfState/SpecificEnthalpy.hpp"
+#include "PointwiseFunctions/Hydro/SpecificEnthalpy.hpp"
 #include "PointwiseFunctions/Hydro/Tags.hpp"  // IWYU pragma: keep
 #include "Utilities/ConstantExpressions.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
@@ -112,7 +112,7 @@ void PrimitiveFromConservative<PrimitiveRecoveryScheme, ThermodynamicDim>::
             .specific_internal_energy_from_density_and_pressure(
                 *rest_mass_density, *pressure);
       })(equation_of_state);
-  *specific_enthalpy = EquationsOfState::specific_enthalpy(
+  *specific_enthalpy = hydro::specific_enthalpy(
       *rest_mass_density, *specific_internal_energy, *pressure);
 }
 }  // namespace ValenciaDivClean
