+++ src/PointwiseFunctions/Hydro/SpecificEnthalpy.cpp
@@ -1,14 +1,14 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "PointwiseFunctions/Hydro/EquationsOfState/SpecificEnthalpy.hpp"
+#include "PointwiseFunctions/Hydro/SpecificEnthalpy.hpp"
 
 #include "DataStructures/DataVector.hpp"  // IWYU pragma: keep
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
 
 /// \cond
-namespace EquationsOfState {
+namespace hydro {
 template <typename DataType>
 Scalar<DataType> specific_enthalpy(
     const Scalar<DataType>& rest_mass_density,
@@ -18,18 +18,17 @@ Scalar<DataType> specific_enthalpy(
                           get(pressure) / get(rest_mass_density)};
 }
 
-}  // namespace EquationsOfState
-
 #define DTYPE(data) BOOST_PP_TUPLE_ELEM(0, data)
 
-#define INSTANTIATE(_, data)                                        \
-  template Scalar<DTYPE(data)> EquationsOfState::specific_enthalpy( \
-      const Scalar<DTYPE(data)>& rest_mass_density,                 \
-      const Scalar<DTYPE(data)>& specific_internal_energy,          \
+#define INSTANTIATE(_, data)                               \
+  template Scalar<DTYPE(data)> specific_enthalpy(          \
+      const Scalar<DTYPE(data)>& rest_mass_density,        \
+      const Scalar<DTYPE(data)>& specific_internal_energy, \
       const Scalar<DTYPE(data)>& pressure) noexcept;
 
 GENERATE_INSTANTIATIONS(INSTANTIATE, (double, DataVector))
 
 #undef DTYPE
 #undef INSTANTIATE
+}  // namespace hydro
 /// \endcond
