+++ tests/Unit/PointwiseFunctions/Hydro/CMakeLists.txt
@@ -7,8 +7,8 @@ set(LIBRARY_SOURCES
   EquationsOfState/Test_DarkEnergyFluid.cpp
   EquationsOfState/Test_IdealFluid.cpp
   EquationsOfState/Test_PolytropicFluid.cpp
-  EquationsOfState/Test_SpecificEnthalpy.cpp
   Test_LorentzFactor.cpp
+  Test_SpecificEnthalpy.cpp
   Test_Tags.cpp
   )
